feat: Add page header with theme toggle and style
Adding the page header provides the title of the site and a styled page for viewers. By default the page displays a dark theme,  users can change between the dark theme and a light theme by clicking the theme toggle button. Jasmine unit tests are included as part of behaviour driven development process.