feat: Split toggle theme into two buttons and moved to side bar
Splitting the theme button into two allows users to select a particular theme, will be easier to add other themes in future. Moving these buttons to the side bar is to allow the sidebar menu to expand without impacting the header. Styles have been added to highlight which theme is active and unit tests updated to reflect new functionality.